requirements.txt
python-dotenv
sounddevice
soundfile
scipy
openai
pydub